canary-tester/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
amqp==5.2.0
annotated-types==0.6.0
asgiref==3.8.1
attrs==23.2.0
billiard==4.2.0
black==24.3.0
blinker==1.7.0
cachetools==5.3.3
celery==5.4.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.3
flake8==7.0.0
Flask==3.0.3
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kombu==5.3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
packaging==24.0
pandas==2.2.1
pathspec==0.12.1
patsy==0.5.6
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.5.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pyarrow==15.0.2
pycodestyle==2.11.1
pydantic==2.7.0
pydantic_core==2.18.1
pydeck==0.8.1b0
pyflakes==3.2.0
Pygments==2.17.2
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
scipy==1.13.0
six==1.16.0
smmap==5.0.1
statsmodels==0.14.1
streamlit==1.33.0
tenacity==8.2.3
toml==0.10.2
toolz==0.12.1
tornado==6.4
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
vine==5.1.0
waitress==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.2
yarl==1.9.4
